fix: secure rooms creation and room lists

+++ back/src/chat/penalty/penalty.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, HttpStatus, HttpException } from '@nestjs/common';
 import { PrismaService } from 'src/prisma/prisma.service';
 import { PenaltyDto } from './Penalty';
 
@@ -6,6 +6,20 @@ import { PenaltyDto } from './Penalty';
 export default class PenaltyService
 {
 	constructor(private readonly prisma: PrismaService) {}
+
+	// checks
+
+	isPenaltyFinished(startPenaltyTime: Date, penaltyDurationInMin: number): boolean
+	{
+		const endPenaltyTime = new Date(startPenaltyTime.getTime() + penaltyDurationInMin * 60000);
+		const currentTime = new Date(Date.now());
+
+		if (endPenaltyTime > currentTime)
+			return (false);
+		return (true);
+	}
+
+	// CRUD
 
 	async createPenalty(penalty: PenaltyDto, authorId: number)
 	{
@@ -33,13 +47,32 @@ export default class PenaltyService
 		});
 	}
 
-	async deletePenalty(penaltyId: number)
+	async getPenalty(penaltyId: number)
 	{
-		await this.prisma.penalty.delete({
+		const penalty = await this.prisma.penalty.findUnique({
 			where: {
 				id: penaltyId
 			}
 		});
+		return (penalty);
+	}
+
+	async deletePenalty(penaltyId: number, userId: number)
+	{
+		const penalty = await this.getPenalty(penaltyId);
+
+		if (penalty && penalty.targetId === userId && this.isPenaltyFinished(penalty.date, penalty.durationInMin))
+		{
+			await this.prisma.penalty.delete({
+				where: {
+					id: penaltyId
+				}
+			});
+		}
+		else
+		{
+			throw new HttpException('Forbidden', HttpStatus.FORBIDDEN);
+		}
 	}
 
 	async deletePenalties(chatId: number, userId: number)
